Refactor for multi-repo configuration
This change prepares the capability of generating a sphinx based documentation using multiple repositories (located outside of the directory of this repo here) as generation source.

NOTE: In order to make the entire multi-repo configuration work (i.e. to make the sphinx build successfully), the  creation/provisioning of additional repositories/content outside of this repo is required now. In particular, a reference (relative path from parent fiolder of this repo) to the '_toc.yml' has to be provided in 'external_toc_path' of conf.py.

Here are the cornerpoints of the change content:
* Utilisation of the 'sphinx_external_toc' extension to conduct YAML based TOC configurations instead of using the toctree:: directive
* Refactoring:
  * The configuration file and static file placesholders has been moved into a dedicated folder
  * The 'SOURCEDIR' as per 'Makefile' has been referenced outside the repo directory
* Cleanup:
  * The Autosar sample .rst files have been removed
  * Minor grammer/typo fixes